Updates age gate process and implements for CURE
