+++ cure/Controllers/KICController.cs
@@ -1,0 +1,35 @@
+﻿using KiCData.Models;
+using KiCData.Services;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Filters;
+
+namespace cure.Controllers
+{
+    public class KICController : Controller
+    {
+        private readonly IConfigurationRoot _configurationRoot;
+        private readonly IUserService _userService;
+        private readonly IHttpContextAccessor _contextAccessor;
+        private readonly KiCdbContext _context;
+        private readonly ICookieService _cookieService;
+
+        public KICController(IConfigurationRoot configurationRoot, IUserService userService, IHttpContextAccessor httpContextAccessor, KiCdbContext kiCdbContext, ICookieService cookieService)
+        {
+            _userService = userService;
+            _configurationRoot = configurationRoot;
+            _context = kiCdbContext;
+            _contextAccessor = httpContextAccessor;
+            _cookieService = cookieService;
+        }
+
+        public override void OnActionExecuting(ActionExecutingContext context)
+        {
+            if (!_cookieService.AgeGateCookieAccepted(_contextAccessor.HttpContext.Request))
+            {
+                context.Result = new RedirectToActionResult("Index", "Home", context.HttpContext.Request.RouteValues);
+            }
+
+            base.OnActionExecuting(context);
+        }
+    }
+}
