+++ cure/Views/Gen/Privacy.cshtml
