+++ cure/Views/Gen/Registration.cshtml
