+++ cure/Views/Gen/Rules.cshtml
