+++ cure/Views/Gen/TicketTypes.cshtml
